Fix bug in getBackupSigningKey
This function was missing a function application at the end, meaning that
its return type was function and not string as it should be. This has
now been fixed.

It took me a bit too long to find this bug so included in this commit is
a small change to make the code a bit less compact by declaring the
return value from getBackupSigningKey to an appropriately named local
variable.

Also improves nomenclature in getBackupSigningKey by explictly
distinguishing between regular private keys and extended private keys.