Utilize signature module from ethereumjs-util
This replaces the use of the `sign` function in `ethereumjs-tx` with the
`ecsign` function from the `ethereumjs-util` package as requested in

We currently chose not to use EIP155 when generating the signatures. To
add this, we would need to know the chain ID for the Ethereum chain we are
signing the transaction. Activating EIP155 would make the generated
signatures replay-resistant on other Ethereum blockchains.